Modify the mutation algorithm
Modify the mutation algorithm so that the test samples generated
by the mutation algorithm have higher grammatical correctness.

Use a mutation algorithm based on grammatical derivation mutation
rules to improve the correctness of Fuzzer's test samples and
increase the efficiency of fuzz testing.

Only nodes that do not have this node type in the node library
are added to subsequent node library.

Set the variable NUMOFNT to the maximum number of nodes of each
type that can subsequently be added to the library.

Signed-off-by: ash2kzzz <ash2kzzz@gmail.com>